+++ python/ray/experimental/workflow/__init__.py
@@ -1,6 +1,6 @@
-from ray.experimental.workflow.api import (step, virtual_actor, run, resume,
+from ray.experimental.workflow.api import (step, virtual_actor, resume,
                                            get_output, get_actor)
 from ray.experimental.workflow.workflow_access import WorkflowExecutionError
 
-__all__ = ("step", "virtual_actor", "run", "resume", "get_output", "get_actor",
+__all__ = ("step", "virtual_actor", "resume", "get_output", "get_actor",
            "WorkflowExecutionError")
