+++ python/ray/experimental/workflow/api.pyi
@@ -77,13 +77,6 @@ def step(function: Callable[[T0, T1, T2, T3, T4, T5, T6, T7, T8, T9], Union[R, W
 def step(function: Callable[..., Union[R, Workflow[R]]]) -> WorkflowStepFunction[R, T0, T1, T2, T3, T4, T5, T6, T7, T8, T9]: ...
 
 @overload
-def run(entry_workflow: Workflow[R]) -> ObjectRef[R]: ...
-@overload
-def run(entry_workflow: Workflow[R], storage: Optional[Union[str, Storage]]) -> ObjectRef[R]: ...
-@overload
-def run(entry_workflow: Workflow[R], storage: Optional[Union[str, Storage]], workflow_id: Optional[str]) -> ObjectRef[R]: ...
-
-@overload
 def resume(workflow_id: str) -> ObjectRef: ...
 @overload
 def resume(workflow_id: str, storage: Optional[Union[str, Storage]]) -> ObjectRef: ...
